fix: Rectify bug of widget views not displaying any data
One must be aware that App Widget layouts are based on RemoteViews, which do not support every kind of layout or view widget. A RemoteViews object (and, consequently, an App Widget) can support the following layout classes:

FrameLayout, LinearLayout, RelativeLayout, GridLayout

And the following widget classes:

AnalogClock, Button, Chronometer, ImageButton, ImageView, ProgressBar, TextView, ViewFlipper, ListView ,GridView,StackView, AdapterViewFlipper

Descendants of these classes are not supported. Also see [What views can i use in an appWidget? ](http://stackoverflow.com/a/16089389/1464571). As such I removed CardViews which are not supported.

This solved the bug that caused widget views to not update with results of the football scores.

Resolves: #1
closes: #1